requirements.txt
flask
pandas
paramiko
jsonify
openpyxl
gunicorn
Flask-SQLAlchemy==3.0.5
Flask-Login
Flask-WTF==1.1.1
python-dotenv==1.0.0
tweepy==4.14.0
vk-api==11.9.9
requests==2.31.0
APScheduler==3.10.4
python-telegram-bot==20.4
APScheduler==3.10.4
python-telegram-bot==20.4
requests-oauthlib==1.3.1
ok-api
instagrapi==1.16.41
yt-dlp==2023.10.13
pytube==15.0.0
moviepy
opencv-python
flask_babel
Werkzeug==2.3.7 